+++ src/main/java/com/sbt/concurrent/SimpleCounter.java
@@ -1,7 +1,7 @@
 package com.sbt.concurrent;
 
 public class SimpleCounter implements Counter{
-    private int i = 0;
+    private long i = 0;
     @Override
     public void increment() {
         i++;
